Move upload logic to a hook; Update toolbar
